etl/pyproject.toml
[project]
name = "etl-pipeline"
version = "0.1.0"
description = "ETL pipeline: MinIO -> PosgreSQL"
requires-python = ">=3.11"

dependencies = [
    "pandas==2.2.2",
    "sqlalchemy==2.0.31",
    "psycopg2-binary==2.9.9",
    "boto3==1.34.131",
    "pyarrow==16.1.0",
    "fastparquet==2024.5.0",
    "python-dotenv==1.0.1"
]

[tool.uv]